% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone (regression practice data)}
\format{
A data frame with 4,177 rows and 9 columns: \code{sex} (factor),
seven physical measurements and the target \code{rings} (integer).
}
\source{
\url{https://archive.ics.uci.edu/dataset/1/abalone}
}
\usage{
abalone
}
\description{
The classic UCI abalone dataset: predict the number of rings (a proxy for
age) of abalone from physical measurements. Included as standard
supervised-learning practice data; it has no causal structure.
}
\examples{
data(abalone)
head(abalone)
}
\keyword{datasets}
