% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalearners.R
\name{ate}
\alias{ate}
\alias{ate.cate_learner}
\title{Average treatment effect of a fitted CATE model}
\usage{
ate(object, ...)

\method{ate}{cate_learner}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted CATE model, e.g. from \code{\link[=s_learner]{s_learner()}}.}

\item{...}{Passed on to methods.}

\item{newdata}{A \code{data.frame} containing at least the covariate columns.}
}
\value{
A single numeric value.
}
\description{
Averages the CATE predictions of a fitted meta-learner over the rows of
\code{newdata}:
\deqn{\widehat{ATE} = \frac{1}{n} \sum_i \hat\tau(x^{(i)})}
}
\examples{
\donttest{
library(mlr3)
data(synth_train)
m <- s_learner(synth_train, outcome = "y", treatment = "t",
               learner = lrn("regr.rpart"))
ate(m, synth_train)
}

}
