% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalmlr-package.R
\docType{package}
\name{causalmlr-package}
\alias{causalmlr}
\alias{causalmlr-package}
\title{causalmlr: Causal Machine Learning with mlr3}
\description{
Estimation and evaluation of average and conditional average treatment
effects (ATE/CATE) from observational data using machine learning, built
on the mlr3 ecosystem.
}
\section{ATE estimators}{

\itemize{
\item \code{\link[=ate_naive]{ate_naive()}} - difference in means (baseline)
\item \code{\link[=ate_ipw]{ate_ipw()}} - inverse propensity weighting
\item \code{\link[=ate_dr]{ate_dr()}} - doubly robust / AIPW
\item \code{\link[=ate_dml]{ate_dml()}} - double machine learning (partially linear model)
}
}

\section{CATE meta-learners}{

\itemize{
\item \code{\link[=s_learner]{s_learner()}}, \code{\link[=t_learner]{t_learner()}}, \code{\link[=x_learner]{x_learner()}}, \code{\link[=dr_learner]{dr_learner()}}
}
}

\section{Evaluation}{

\itemize{
\item \code{\link[=eps_ate]{eps_ate()}} - absolute ATE error (needs ground truth)
\item \code{\link[=pehe]{pehe()}} - precision in estimating heterogeneous effects (needs ground truth)
\item \code{\link[=r_loss]{r_loss()}} with \code{\link[=rloss_nuisance]{rloss_nuisance()}} - observable score for model
selection and hyperparameter tuning on real data
}

All estimators accept arbitrary mlr3 learners as nuisance models, so
anything from \code{mlr3::lrn("regr.lm")} to a tuned \code{AutoTuner} can be
plugged in.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmachlanski/causalmlr}
  \item Report bugs at \url{https://github.com/dmachlanski/causalmlr/issues}
}

}
\author{
\strong{Maintainer}: Damian Machlanski \email{d.machlanski@ed.ac.uk}

Authors:
\itemize{
  \item Damian Machlanski \email{d.machlanski@ed.ac.uk}
}

}
\keyword{internal}
