% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{eps_ate}
\alias{eps_ate}
\title{Absolute error of an ATE estimate}
\usage{
eps_ate(ate_true, ate_pred)
}
\arguments{
\item{ate_true}{True ATE (a single number).}

\item{ate_pred}{Estimated ATE. Either a number or a \code{"causalmlr_ate"}
object as returned by \code{\link[=ate_naive]{ate_naive()}} and friends.}
}
\value{
A single non-negative number.
}
\description{
Measures the absolute difference between an estimated and the true average
treatment effect:
\deqn{\epsilon_{ATE} = | \widehat{ATE} - ATE |}
Requires the true ATE, so it is only usable with (semi-)synthetic
benchmark data such as \link{sodium} or \link{ihdp_train}.
}
\examples{
data(sodium)
est <- ate_naive(sodium, outcome = "bp", treatment = "sodium")
eps_ate(1.05, est)

}
\seealso{
\code{\link[=pehe]{pehe()}}, \code{\link[=r_loss]{r_loss()}}
}
