% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housing}
\alias{housing}
\title{Boston housing (regression practice data)}
\format{
A data frame with 506 rows and 14 columns.
}
\usage{
housing
}
\description{
The Boston housing dataset (Harrison & Rubinfeld, 1978): predict median
house value (\code{MEDV}) from neighbourhood characteristics. Included as
standard supervised-learning practice data. Note that this dataset has
known ethical concerns (the \code{B} column encodes a racial statistic) and is
provided for teaching purposes only.
}
\examples{
data(housing)
head(housing)
}
\keyword{datasets}
