% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobs_test}
\alias{jobs_test}
\title{Jobs benchmark (test set)}
\format{
A data frame with 321 rows and 20 columns; same columns as
\link{jobs_train}.
}
\usage{
jobs_test
}
\description{
Held-out portion of the Jobs benchmark; see \link{jobs_train} for details.
}
\examples{
data(jobs_test)
head(jobs_test)
}
\references{
LaLonde, R. J. (1986). Evaluating the econometric evaluations of training
programs with experimental data. \emph{The American Economic Review}, 76(4),
604-620.

Shalit, U., Johansson, F. D., & Sontag, D. (2017). Estimating individual
treatment effect: generalization bounds and algorithms. \emph{Proceedings of
the 34th International Conference on Machine Learning}.
}
\keyword{datasets}
