% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pehe}
\alias{pehe}
\title{Precision in estimation of heterogeneous effects (PEHE)}
\usage{
pehe(tau_true, tau_pred)
}
\arguments{
\item{tau_true}{Numeric vector of true CATEs.}

\item{tau_pred}{Numeric vector of predicted CATEs (same length).}
}
\value{
A single non-negative number.
}
\description{
The root mean squared error between true and predicted conditional
average treatment effects:
\deqn{PEHE = \sqrt{\frac{1}{n} \sum_i (\tau(x^{(i)}) -
  \hat\tau(x^{(i)}))^2}}
Requires the true CATEs, so it is only usable with (semi-)synthetic
benchmark data such as \link{synth_test} or \link{ihdp_test}.
}
\examples{
pehe(c(1, 2, 3), c(1.1, 1.8, 3.2))

}
\references{
Hill, J. L. (2011). Bayesian nonparametric modeling for causal inference.
\emph{Journal of Computational and Graphical Statistics}, 20(1), 217-240.
}
\seealso{
\code{\link[=eps_ate]{eps_ate()}}, \code{\link[=r_loss]{r_loss()}}
}
