% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sodium}
\alias{sodium}
\title{Sodium intake and blood pressure (synthetic)}
\format{
A data frame with 10,000 rows and 3 columns:
\describe{
\item{age}{Age in years (confounder).}
\item{sodium}{Binary treatment: high sodium intake (1) or not (0).}
\item{bp}{Systolic blood pressure (outcome).}
}
}
\usage{
sodium
}
\description{
A simplified simulation of the effect of sodium intake on blood pressure,
proposed by Luque-Fernandez et al. (2019). Age confounds both the
treatment (high sodium intake) and the outcome (blood pressure). The true
ATE used in the data generating process is \strong{1.05}.
}
\examples{
data(sodium)
ate_naive(sodium, outcome = "bp", treatment = "sodium")
}
\references{
Luque-Fernandez, M. A., Schomaker, M., Redondo-Sanchez, D., Jose
Sanchez Perez, M., Vaidya, A., & Schnitzer, M. E. (2019). Educational
Note: Paradoxical collider effect in the analysis of non-communicable
disease epidemiological data. \emph{International Journal of Epidemiology},
48(2), 640-653.
}
\keyword{datasets}
